requirements/base.txt
# Django (4.2 LTS - 안정적, 모든 라이브러리 호환)
Django==4.2.11
django-environ==0.11.2
django-cors-headers==4.3.1

# Django REST Framework
djangorestframework==3.14.0
drf-yasg==1.21.7

# Stripe (결제)
stripe==8.0.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL

# OpenAI
openai==1.12.0
httpx==0.24.1  # OpenAI 호환성 (proxies 이슈 해결)

# 데이터 수집 (미국 주식 - EDGAR API만 사용)
sec-edgar-api==1.0.0  # EDGAR API (SEC 공식 데이터, 재무제표 전체)
pyrate-limiter==2.10.0  # sec-edgar-api 의존성 (버전 고정 필수)
sec-cik-mapper==2.1.0  # Ticker ↔ CIK 매핑 (SEC 공식 데이터 포함)
finance-datareader==0.9.50  # S&P 500, NASDAQ, NYSE 종목 리스트
pandas==2.2.0  # 데이터 처리
requests==2.31.0  # HTTP 요청
beautifulsoup4==4.12.3  # 웹 크롤링용 (백업)
lxml==5.1.0  # HTML 파싱

# Celery (백그라운드 작업)
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1

# 유틸리티
python-dateutil==2.8.2
numpy==1.26.3

# 파일 처리
PyPDF2==3.0.1  # PDF 파싱
lxml==5.1.0

# 기타
python-dotenv==1.0.1
Pillow==10.2.0

